# harness-docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 职达 Sprint 1—3 API
  version: 0.3.0-draft
  description: |
    职达 Sprint 1—3 核心闭环 API 评审草案。

    本文档由 AI 辅助整理，必须由三名成员人工审阅和修改后才能标记为 Accepted。
    当前阶段使用固定演示用户，不提供注册登录和文件上传。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/v1/job-descriptions:
    get:
      tags: [Job]
      summary: 查询当前演示用户的 JD 列表
      operationId: listJobDescriptions
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescriptionListResponse"

    post:
      tags: [Job]
      summary: 创建原始 JD
      operationId: createJobDescription
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobDescriptionRequest"
      responses:
        "201":
          description: 创建成功，尚未解析
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescriptionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/job-descriptions/{jobDescriptionId}:
    get:
      tags: [Job]
      summary: 查询单个 JD
      operationId: getJobDescription
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/JobDescriptionId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescriptionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/job-descriptions/{jobDescriptionId}/parse:
    post:
      tags: [Job]
      summary: 使用千问 Max 解析 JD
      description: |
        模型只负责把原始 JD 转换为符合 Schema 的岗位要求。解析结果是匹配算法的输入，
        不能直接成为岗位匹配分数。
      operationId: parseJobDescription
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/JobDescriptionId"
      responses:
        "200":
          description: 解析成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescriptionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/AiOutputInvalid"
        "503":
          $ref: "#/components/responses/AiUnavailable"
...
  /api/v1/resume-versions/{versionId}/job-matches:
    post:
      tags: [Matching]
      summary: 计算简历版本与岗位的匹配结果
      description: |
        AI 禁飞区接口。千问解析后的岗位标签只能作为输入，最终相似度和排序规则由成员 2
        手写的确定性程序完成。

        接口采用嵌套资源路径：匹配是简历版本的子资源。
        相同输入重复请求返回已有结果（幂等）。
      operationId: createJobMatch
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/VersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequest"
      responses:
        "200":
          description: 匹配成功；相同输入和算法版本允许返回已存在结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleViolation"
  /api/v1/job-matches/{matchId}:
    get:
      tags: [Matching]
      summary: 查询已有匹配结果
      description: 根据匹配 ID 获取历史匹配记录。
      operationId: getJobMatch
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/interview-plans:
    post:
      tags: [Interview]
      summary: 创建一次多轮模拟面试计划
      description: |
        创建 plan 级面试容器，并为每位面试官创建一个 session。计划层只组织轮次，
        不负责状态机转换。
      operationId: createInterviewPlan
      responses:
        "201":
          description: 创建成功
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/interview-plans/{planId}:
    get:
      tags: [Interview]
      summary: 查询一次多轮面试计划详情
      operationId: getInterviewPlan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: 查询成功
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Interview]
      summary: 从历史列表隐藏一次多轮面试计划
      description: |
        使用 deleted_at 软删除计划历史入口，不物理删除面试问答数据。
      operationId: hideInterviewPlan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: 隐藏成功
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/interview-plans/{planId}/summary:
    post:
      tags: [Interview]
      summary: 生成一次多轮面试的整次复盘
      description: |
        只有 plan 下所有 session 均为 COMPLETED 时才允许生成。生成结果写入
        interview_plans.summary_json，并触发成长趋势快照写入。
      operationId: generateInterviewPlanSummary
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: 复盘生成成功
        "400":
          $ref: "#/components/responses/BusinessRuleViolation"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: 客户端可传入的请求关联 ID；缺失时由服务端生成
      schema:
        type: string
        maxLength: 64
...
    VersionId:
      name: versionId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    JobDescriptionId:
      name: jobDescriptionId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
    MatchId:
      name: matchId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不属于当前演示用户
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BusinessRuleViolation:
      description: 业务规则校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AiOutputInvalid:
      description: AI 返回内容未通过结构化校验
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AiUnavailable:
      description: 模型超时、限流或供应商暂时不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Id:
      type: integer
      format: int64
      minimum: 1
    Timestamp:
      type: string
      format: date-time
...
    ParseStatus:
      type: string
      enum: [pending, succeeded, failed]
...
    ApiMeta:
      type: object
      additionalProperties: false
      required: [requestId]
      properties:
        requestId:
          type: string
...
    CreateJobDescriptionRequest:
      type: object
      additionalProperties: false
      required: [jobTitle, rawText]
      properties:
        jobTitle:
          type: string
          minLength: 1
          maxLength: 200
        companyName:
          type: [string, "null"]
          maxLength: 200
        rawText:
          type: string
          minLength: 20
          maxLength: 50000
    ParsedJobDescription:
      type: object
      additionalProperties: false
      required:
        - requiredSkills
        - preferredSkills
        - responsibilities
        - experienceRequirements
        - educationRequirements
      properties:
        requiredSkills:
          type: array
          uniqueItems: true
          items:
            type: string
        preferredSkills:
          type: array
          uniqueItems: true
          items:
            type: string
        responsibilities:
          type: array
          items:
            type: string
        experienceRequirements:
          type: array
          items:
            type: string
        educationRequirements:
          type: array
          items:
            type: string
    JobDescription:
      type: object
      additionalProperties: false
      required:
        [id, jobTitle, rawText, parseStatus, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        jobTitle:
          type: string
        companyName:
          type: [string, "null"]
        rawText:
          type: string
        parsed:
          oneOf:
            - $ref: "#/components/schemas/ParsedJobDescription"
            - type: "null"
        parseStatus:
          $ref: "#/components/schemas/ParseStatus"
        promptVersion:
          type: [string, "null"]
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...
    MatchRequest:
      type: object
      additionalProperties: false
      required: [jobDescriptionId]
      properties:
        jobDescriptionId:
          $ref: "#/components/schemas/Id"
    MatchDetails:
      type: object
      additionalProperties: false
      properties:
        requiredCoverage:
          type: integer
          minimum: 0
          maximum: 100
        preferredCoverage:
          type: integer
          minimum: 0
          maximum: 100
        experienceCoverage:
          type: integer
          minimum: 0
          maximum: 100
        educationMatch:
          type: boolean
        responsibilityCoverage:
          type: integer
          minimum: 0
          maximum: 100
        matchedItems:
          type: array
          items:
            type: string
        missingItems:
          type: array
          items:
            type: string
        unknownItems:
          type: array
          items:
            type: string
        dimensionScores:
          type: object
          additionalProperties:
            type: integer
        aliasMatches:
          type: object
          additionalProperties:
            type: string
    MatchResponse:
      type: object
      additionalProperties: false
      required:
        - matchScore
        - details
      properties:
        matchScore:
          type: integer
          minimum: 0
          maximum: 100
        details:
          $ref: "#/components/schemas/MatchDetails"
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: 稳定的机器可读错误码
        message:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    JobDescriptionResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/JobDescription"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    JobDescriptionListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/JobDescription"
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/ApiMeta"